requirements.txt
agentql>=0.1.0
python-dotenv>=0.19.0
flask>=2.0.0
geopy>=2.3.0
pandas>=1.5.0
